--- packages/plugin-shiki/src/shiki/transformers/index.ts
@@ -1,1 +1,0 @@
-export * from './line-number';
