+++ packages/plugin-shiki/src/transformers/index.ts
@@ -1,0 +1,5 @@
+export {
+  transformerLineNumber,
+  SHIKI_TRANSFORMER_LINE_NUMBER,
+} from './line-number';
+export { transformerAddTitle, SHIKI_TRANSFORMER_ADD_TITLE } from './add-title';
